refactor HTMLRenderer to use shared parsing
